fragmentation complete
