+++ proposals/dao/v2Phase1-part2.ts
@@ -1,3 +1,24 @@
+import '@nomiclabs/hardhat-ethers';
+import hre, { ethers } from 'hardhat';
+import {
+  SetupUpgradeFunc,
+  ValidateUpgradeFunc,
+  RunUpgradeFunc,
+  TeardownUpgradeFunc,
+  NamedContracts
+} from '@custom-types/types';
+
+import chai, { expect } from 'chai';
+import CBN from 'chai-bn';
+import { CollateralizationOracleKeeper, Core, EthBondingCurve, PCVEquityMinter, RatioPCVController, TribeReserveStabilizer, UniswapPCVDeposit } from '@custom-types/contracts';
+import { getImpersonatedSigner } from "@test/helpers";
+
+const toBN = ethers.BigNumber.from;
+
+before(() => {
+  chai.use(CBN(ethers.BigNumber));
+});
+
 /*
 
 V2 Phase 1 Upgrade
@@ -43,3 +64,19 @@ DAO ACTIONS:
 (no actions by the dao)
 
 */
+
+export const setup: SetupUpgradeFunc = async (addresses, oldContracts, contracts, logging) => {
+    logging && console.log(`V2-Phase1 part 2 has no DAO setup actions.`)
+};
+  
+export const run: RunUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    logging && console.log(`V2-Phase1 part 2 has no DAO run actions.`)
+};
+
+export const teardown: TeardownUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    logging && console.log(`V2-Phase1 part 2 has no DAO teardown actions.`)
+}
+
+export const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    logging && console.log(`V2-Phase1 part 2 has no DAO validate actions.`)
+}
