+++ proposals/dao/v2Phase1-part3.ts
@@ -1,3 +1,21 @@
+import '@nomiclabs/hardhat-ethers';
+import hre, { ethers } from 'hardhat';
+import {
+  SetupUpgradeFunc,
+  ValidateUpgradeFunc,
+  RunUpgradeFunc,
+  TeardownUpgradeFunc
+} from '@custom-types/types';
+
+import chai, { expect } from 'chai';
+import CBN from 'chai-bn';
+
+const toBN = ethers.BigNumber.from;
+
+before(() => {
+  chai.use(CBN(ethers.BigNumber));
+});
+
 /*
 
 V2 Phase 1 Upgrade
@@ -36,3 +54,113 @@ DAO ACTIONS:
 6. Seed TRIBE to LBP Swapper
 
 */
+
+export const setup: SetupUpgradeFunc = async (addresses, oldContracts, contracts, logging) => {
+    const { timelock } = addresses;
+  
+    await hre.network.provider.request({
+      method: 'hardhat_impersonateAccount',
+      params: [timelock]
+    });
+};
+  
+export const run: RunUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    const { timelock: timelockAddress, rariPool19DpiPCVDeposit: rariPool19DpiPCVDepositAddress } = addresses;
+  
+    const {
+      dpiUniswapPCVDeposit,
+      uniswapPCVDeposit,
+      bondingCurve,
+      dpiBondingCurve,
+      tribeReserveStabilizer,
+      ratioPCVController,
+      feiTribeLBPSwapper,
+      core,
+      tribe
+    } = contracts;
+  
+    const oldRatioPCVController = oldContracts.ratioPCVController;
+
+    // special role
+    // check via tribe contract
+    logging && console.log('Transferring TRIBE Minter role to TribeReserveStabilizer');
+
+    await hre.network.provider.request({
+        method: 'hardhat_impersonateAccount',
+        params: [timelockAddress]
+    });
+
+    const timelockSigner = await ethers.getSigner(timelockAddress);
+
+    await tribe.connect(timelockSigner).setMinter(tribeReserveStabilizer.address);
+
+    logging && console.log('Setting mint cap.');
+    await bondingCurve.setMintCap(ethers.constants.MaxUint256);
+
+    logging && console.log(`Withdrawing ratio from old uniswap pcv deposit to new.`);
+    await oldRatioPCVController.withdrawRatio(oldContracts.uniswapPCVDeposit.address, uniswapPCVDeposit.address, '10000'); // move 100% of PCV from old -> new
+
+    logging && console.log(`Withdrawing ratio from old dpi uniswap pcv deposit to new.`)
+    await ratioPCVController.withdrawRatio(
+        oldContracts.dpiUniswapPCVDeposit.address,
+        dpiUniswapPCVDeposit.address,
+        '10000'
+    ); // move 100% of PCV from old -> new
+
+    logging && console.log(`Allocating Tribe...`);
+    await core.allocateTribe(feiTribeLBPSwapper.address, toBN('1000000').mul(ethers.constants.WeiPerEther));
+
+    logging && console.log(`Setting allocation...`);
+    await dpiBondingCurve.setAllocation([dpiUniswapPCVDeposit.address, rariPool19DpiPCVDepositAddress], ['9000', '1000']);
+}
+
+export const teardown: TeardownUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    const core = contracts.core;
+    const { uniswapPCVController } = addresses;
+    const { ratioPCVController } = oldContracts;
+  
+    await core.revokePCVController(ratioPCVController.address);
+    await core.revokePCVController(uniswapPCVController);
+    await core.revokeMinter(uniswapPCVController);
+  
+    // Deposit Uni and DPI
+    await contracts.dpiUniswapPCVDeposit.deposit();
+    await contracts.uniswapPCVDeposit.deposit();
+}
+
+export const validate: ValidateUpgradeFunc = async (addresses, oldContracts, contracts, logging = false) => {
+    const { dai: daiAddress, dpi: dpiAddress, rai: raiAddress, fei: feiAddress, weth: wethAddress } = addresses;
+    const {
+      collateralizationOracle,
+      collateralizationOracleWrapperImpl,
+      collateralizationOracleWrapper,
+      core,
+      proxyAdmin,
+      feiTribeLBPSwapper
+    } = contracts;
+
+    const pcvStatsCurrent = await collateralizationOracleWrapper.pcvStatsCurrent();
+    const pcvStats = await collateralizationOracle.pcvStats();
+  
+    expect(pcvStatsCurrent[0].toString()).to.be.equal(pcvStats[0].toString());
+    expect(pcvStatsCurrent[1].toString()).to.be.equal(pcvStats[1].toString());
+    expect(pcvStatsCurrent[2].toString()).to.be.equal(pcvStats[2].toString());
+    expect(pcvStatsCurrent[3].toString()).to.be.equal(pcvStats[3].toString());
+  
+    await collateralizationOracleWrapper.update();
+  
+    expect((await collateralizationOracle.getTokensInPcv()).length).to.be.equal(6);
+    expect((await collateralizationOracle.getDepositsForToken(daiAddress)).length).to.be.equal(2);
+    expect((await collateralizationOracle.getDepositsForToken(dpiAddress)).length).to.be.equal(3);
+    expect((await collateralizationOracle.getDepositsForToken(raiAddress)).length).to.be.equal(3);
+    expect((await collateralizationOracle.getDepositsForToken(wethAddress)).length).to.be.equal(6);
+    expect((await collateralizationOracle.getDepositsForToken(feiAddress)).length).to.be.equal(11);
+  
+    expect(await feiTribeLBPSwapper.CONTRACT_ADMIN_ROLE()).to.be.not.equal(await core.GOVERN_ROLE());
+  
+    expect(await proxyAdmin.getProxyImplementation(collateralizationOracleWrapper.address)).to.be.equal(
+      collateralizationOracleWrapperImpl.address
+    );
+    
+    expect(await proxyAdmin.getProxyAdmin(collateralizationOracleWrapper.address)).to.be.equal(proxyAdmin.address);
+}
